+++ docs/docs-components/siteIndex.js
@@ -260,10 +260,12 @@ const siteIndex: $ReadOnlyArray<siteIndexType> = [
         pages: [
           'About international design',
           'Icon localization',
+          'Pluralization',
           {
             sectionName: 'RTL guidelines',
             pages: ['RTL overview', 'Layout and text direction', 'Iconography', 'Typography'],
           },
+          'Number localization',
         ],
       },
       'Layout',